Add columns Campaign, Accepted and Rerouted
Add requested columns to migration Daily performances report

+++ database/migrations/2019_07_04_081542_add_active_field_to_supervisors_table.php
@@ -25,8 +25,10 @@ class AddActiveFieldToSupervisorsTable extends Migration
      */
     public function down()
     {
-        Schema::table('supervisors', function (Blueprint $table) {
-            $table->dropColumn('active');
-        });
+        if (Schema::hasColumn('supervisors', 'active')) {
+            Schema::table('supervisors', function (Blueprint $table) {
+                $table->dropColumn('active');
+            }); 
+        }        
     }
 }
