+++ database/migrations/2019_07_11_103911_add_file_name_field_to_performances_table.php
@@ -21,8 +21,10 @@ class AddFileNameFieldToPerformancesTable extends Migration
      */
     public function down()
     {
-        Schema::table('performances', function (Blueprint $table) {
-            $table->dropColumn('file_name');
-        });
+        if (Schema::hasColumn('performances', 'file_name')) {
+            Schema::table('performances', function (Blueprint $table) {
+                $table->dropColumn('file_name');
+            });
+        }        
     }
 }
