+++ database/migrations/2019_07_20_084853_add_slug_field_to_login_names_table.php
@@ -21,8 +21,11 @@ class AddSlugFieldToLoginNamesTable extends Migration
      */
     public function down()
     {
-        Schema::table('login_names', function (Blueprint $table) {
-            $table->dropColumn('slug');
-        });
+        if (Schema::hasColumn('login_names', 'slug')) {
+            Schema::table('login_names', function (Blueprint $table) {
+                $table->dropColumn('slug');
+            });
+        }
+        
     }
 }
